wip: refining the ostd spec and draft implementation
